# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Auspac Documentation",
        "description": "Project Inventory Auspac Laravel 12 Documentation",
        "contact": {
            "email": "example@email.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/stockbuys": {
            "get": {
                "tags": [
                    "StockBuys"
                ],
                "summary": "Get all stock buys",
                "description": "Retrieve all stock buy records, including the related product information.",
                "operationId": "10e9abba85f1f449c355f6a4867b034f",
                "responses": {
                    "200": {
                        "description": "List of stock buys retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/StockBuy"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "StockBuys"
                ],
                "summary": "Create a new stock buy",
                "description": "Registers a new stock purchase, increasing the stock quantity for a specific product.",
                "operationId": "48b859bc211264aca380a690a15d1ac8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product",
                                    "amount",
                                    "date",
                                    "description"
                                ],
                                "properties": {
                                    "product": {
                                        "$ref": "#/components/schemas/Product"
                                    },
                                    "amount": {
                                        "description": "Quantity of units purchased.",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "date": {
                                        "description": "Purchase date.",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-05"
                                    },
                                    "description": {
                                        "description": "Short description of the purchase.",
                                        "type": "string",
                                        "example": "Purchase #001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "StockBuy created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StockBuy"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "Unexpected server error"
                    }
                }
            }
        },
        "/api/stockbuys/{id}": {
            "get": {
                "tags": [
                    "StockBuys"
                ],
                "summary": "Get a specific stock buy",
                "description": "Retrieve a single stock buy record by its ID, including the associated product details.",
                "operationId": "b6c8ac789db3097a4921ac883114d595",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the stock buy record",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "StockBuy found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StockBuy"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "StockBuy not found"
                    },
                    "500": {
                        "description": "Unexpected server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "StockBuys"
                ],
                "summary": "Update a stock buy",
                "description": "Update an existing stock buy by its ID, adjusting the stock levels of the associated product accordingly.",
                "operationId": "904f8798396d65136cacb67bdf577b51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the stock buy record to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StockBuy"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "StockBuy updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StockBuy"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "StockBuy not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "Unexpected server error"
                    }
                }
            }
        },
        "/api/stock-buys/{id}": {
            "delete": {
                "tags": [
                    "StockBuys"
                ],
                "summary": "Delete a stock purchase",
                "description": "Delete an existing stock purchase. This will reverse the stock increase applied when the purchase was created.",
                "operationId": "914332a30a0e9ba320033db40414b0bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the stock purchase",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "StockBuy deleted successfully"
                    },
                    "404": {
                        "description": "StockBuy not found"
                    },
                    "500": {
                        "description": "Unexpected server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Brand": {
                "title": "Brand",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "BS"
                    }
                },
                "type": "object"
            },
...
            "Coin": {
                "title": "Coin",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "USD"
                    }
                },
                "type": "object"
            },
            "Measure": {
                "title": "Measure",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "UNIT"
                    }
                },
                "type": "object"
            },
            "Product": {
                "title": "Product",
                "description": "Represents a product with its pricing, dimensions, and associations.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "P001"
                    },
                    "supplier_cost_price": {
                        "type": "number",
                        "format": "float",
                        "example": 100.5
                    },
                    "supplier_coin": {
                        "$ref": "#/components/schemas/Coin"
                    },
                    "landing_cost_price": {
                        "type": "number",
                        "format": "float",
                        "example": 120.75
                    },
                    "landing_coin": {
                        "$ref": "#/components/schemas/Coin"
                    },
                    "retail_price": {
                        "type": "number",
                        "format": "float",
                        "example": 200
                    },
                    "promotional_price": {
                        "type": "number",
                        "format": "float",
                        "example": 180
                    },
                    "stock": {
                        "type": "integer",
                        "example": 50
                    },
                    "measure": {
                        "$ref": "#/components/schemas/Measure"
                    },
                    "serial_tracking": {
                        "type": "string",
                        "example": "SER001"
                    },
                    "dimension_size": {
                        "type": "string",
                        "example": "10x10x10"
                    },
                    "dimension_weight": {
                        "type": "integer",
                        "example": 5
                    },
                    "sub_brand": {
                        "$ref": "#/components/schemas/SubBrand"
                    },
                    "supplier": {
                        "$ref": "#/components/schemas/Supplier"
                    }
                },
                "type": "object"
            },
            "StockBuy": {
                "title": "StockBuy",
                "description": "Represents a purchase record that increases the stock of a product.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    },
                    "amount": {
                        "description": "Quantity of units purchased.",
                        "type": "integer",
                        "example": 10
                    },
                    "date": {
                        "description": "Date when the purchase occurred.",
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-05"
                    },
                    "description": {
                        "description": "Short description or reference for this purchase.",
                        "type": "string",
                        "example": "Purchase #001"
                    }
                },
                "type": "object"
            },
...
            "SubBrand": {
                "title": "SubBrand",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "ACME_ECO"
                    },
                    "brand": {
                        "$ref": "#/components/schemas/Brand"
                    }
                },
                "type": "object"
            },
            "Supplier": {
                "title": "Supplier",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Supplier 1"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}